# lib/units/api/swagger/api_v1.yaml
swagger: "2.0"
info:
  version: "2.0.0"
  title: Smartphone Test Farm
  description: Control and manages real Smartphone devices from browser and restful apis
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: STF Team
    email: contact@openstf.io
    url: http://openstf.io/
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /devices:
    x-swagger-router-controller: devices
    get:
      summary: Device List
      description: The devices endpoint return list of all the STF devices including Disconnected and Offline
      operationId: getDevices
      tags:
        - devices
      parameters:
        - name: fields
          in: query
          description: Fields query parameter takes a comma seperated list of fields. Only listed field will be return in response
          required: false
          type: string
        - name: filter
          in: query
          description: |
            An optional query parameter to filter the results. You can filter the response by model, name, serial, operator, version, manufacturer, sdk, imei, and status. The filter values must exactly match. For example to retrieve the devices for the manufacturer "SONY", use /api/v1/devices/?filter=manufacturer:SONY for the filter query parameter.
            Use the ampersand (&) to combine filters. For example, to return both operators "KDDI" and "DOCOMO" of the devices, combine these filters:
            filter=operator:KDDI&filter=operator:DOCOMO
          required: false
          type: string
      responses:
        "200":
          description: List of Devices
          schema:
            $ref: "#/definitions/DeviceListResponse"
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - accessTokenAuth: []
  /devices/{serial}:
    x-swagger-router-controller: devices
    get:
      summary: Device Information
      description: The device enpoint return information about a single device
      operationId: getDeviceBySerial
      tags:
        - devices
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
        - name: fields
          in: query
          description: Fields query parameter takes a comma seperated list of fields. Only listed field will be return in response
          required: false
          type: string
      responses:
        "200":
          description: Device Information
          schema:
            $ref: "#/definitions/DeviceResponse"
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - accessTokenAuth: []
...
definitions:
...
  DeviceListResponse:
    required:
      - devices
    properties:
      devices:
        type: array
        items:
          type: object
  DeviceResponse:
    required:
      - device
    properties:
      device:
        type: object
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
securityDefinitions:
  accessTokenAuth:
    type: apiKey
    name: authorization
    in: header